First commit from desktop env in India. Bug fix for picking image in visual ingredient add view.

+++ app/src/main/java/com/example/sushrut/recipedemo/ImageProcessor.java
@@ -59,7 +59,10 @@ public class ImageProcessor {
                 }
             }
         }
-        CommonUtils.DeleteCameraImage(ci.getContext(),ci.getImageUri());
+
+        if(ci.isImageToBeDeleted())
+            //Image is deleted only if captured from camera. This was cause for the security exception.
+            CommonUtils.DeleteCameraImage(ci.getContext(),ci.getImageUri());
         return detectedShape;
     }
     private List<MatOfPoint> GetContourMatPointList(){
